+++ man-roxygen/param-ellipsis-select-filter.R
@@ -1,0 +1,2 @@
+#' @param ... Additional argument for \link{readLAS} to reduce the amount of data loaded (only with a
+#' \code{LAScatalog} object)