+++ man-roxygen/section-supported-option-tree_detection.R
@@ -1,0 +1,10 @@
+#' @section Supported processing options for a LAScatalog:
+#' \itemize{
+#' \item \strong{cores}: See \link{LAScatalog-class}.
+#' \item \strong{progress}: See \link{LAScatalog-class}.
+#' \item \strong{stop_early}: See \link{LAScatalog-class}.
+#' \item \strong{by_file}: See \link{LAScatalog-class}.
+#' \item \strong{tiling_size}: See \link{LAScatalog-class}.
+#' \item \strong{buffer*}: Requiered to find the tree without edge effet. Something twice bigger than
+#' any possible tree might be a good value. See \link{LAScatalog-class}
+#' }